break up logic in Verifier
